Add files via upload
In high income countries, more than 70% of breast cancer patients are diagnosed in stages I and II. In low- and middle-income countries 20%-50%of patients are diagnosed in these stages. Mortality rates are higher in developing countries. Studies show there is an association between the advanced and delays in getting treatment. To improve prognosis and chances of survival, early diagnosis and timely clinical treatment to patients is important. Accurate classification of tumours to either benign or malignant is important to avoid undergoing unnecessary treatment. Early detection will also help one to know the type of cancer one has(invasive lobular carcinoma or Invasive ductal carcinoma).This subject has led to a lot of research. Machine learning is a widely used methodology in breast cancer classification and forecast modelling because of critical feature detection in complex datasets.